Move to windows.py
